有休・振休勤怠情報削除時に日数戻す処理追加

+++ SmartAttendance/src/main/java/com/example/demo/controller/EditKintaiController.java
@@ -15,16 +15,17 @@ import lombok.RequiredArgsConstructor;
 @Controller
 @RequiredArgsConstructor
 public class EditKintaiController {
-	
+
 	private final MainController mainController;
 	private final EditKintaiService service;
-	
+
 	LoginUser lu = new LoginUser();
 
 	@PostMapping("/edit-kintai")
-	public String editKintai(@ModelAttribute LoginUser loginUser, @ModelAttribute EditKintaiForm editKintaiForm, Model model) {
+	public String editKintai(@ModelAttribute LoginUser loginUser, @ModelAttribute EditKintaiForm editKintaiForm,
+			Model model) {
 		lu = loginUser;
-		
+
 		System.out.println(lu);
 		System.out.println(editKintaiForm);
 		model.addAttribute("editKintaiForm", editKintaiForm);
@@ -33,7 +34,8 @@ public class EditKintaiController {
 	}
 
 	@PostMapping({ "/delete-kintai", "/update-kintai" })
-	public String processKintai(@ModelAttribute("editKintaiForm") EditKintaiForm editKintaiForm, @RequestParam("operation") String operation, @ModelAttribute LoginUser loginUser, Model model) {
+	public String processKintai(@ModelAttribute("editKintaiForm") EditKintaiForm editKintaiForm,
+			@RequestParam("operation") String operation, @ModelAttribute LoginUser loginUser, Model model) {
 		model.addAttribute("editKintaiForm", editKintaiForm);
 		model.addAttribute("loginUser", loginUser);
 		if ("delete".equals(operation)) {
@@ -46,16 +48,36 @@ public class EditKintaiController {
 			return "confirm-edit-kintai";
 		}
 	}
-	
+
 	@PostMapping("/complete-update-kintai")
 	public String completeUpdateKintai(@ModelAttribute EditKintaiForm form, Model model) {
 		service.updateKintai(form);
 		return "redirect:/complete";
 	}
-	
+
 	@PostMapping("/complete-delete-kintai")
 	public String completeDeleteKintai(@ModelAttribute EditKintaiForm form, Model model) {
-		service.deleteKintai(form);
+		if (form.getStart_H() <= 8 && 22 <= form.getEnd_H()) {
+			if (form.getStart_H() > form.getEnd_H()) {
+				model.addAttribute("msg", "始業・終業時刻が不適切です。");
+				return "regist-kintai";
+			} else if (form.getStart_H() >= form.getEnd_H() && form.getStart_M() >= form.getEnd_M()) {
+				model.addAttribute("msg", "始業・終業時刻が不適切です。");
+				return "regist-kintai";
+			}
+			if (form.getTotal_workingTime() >= 4.00 && form.getBreakTime() < 1) {
+				model.addAttribute("msg", "最低一時間以上の休憩時間を入力してください。");
+				return "regist-kintai";
+			}
+			if (form.getTotal_workingTime() <= 0) {
+				model.addAttribute("msg", "実労働時間が入力されていません。");
+				return "regist-kintai";
+			}
+			service.deleteKintai(form);
+		} else if (form.getStart_H() <= 0 && form.getEnd_H() <= 0) {
+			service.checkRecordDate(form);
+			service.deleteKintai(form);
+		} 
 		return "redirect:/complete";
 	}
 }
